refactored save_picture into utils
