+++ flask_blog/users/views/account.py
@@ -1,31 +1,10 @@
-import os
-import secrets
-
 from flask import render_template, url_for, flash, redirect, request, current_app
 from flask_login import login_required, current_user
-from PIL import Image
 
 from extensions import db
 from flask_blog.users.forms.update_account_form import UpdateAccountForm
+from flask_blog.users.utils.save_picture import save_picture
 from flask_blog.users.views import users_blueprint
-
-BASE_DIR = os.path.dirname(os.path.dirname(os.path.dirname(__file__)))
-
-image_dir = os.path.join(BASE_DIR, "static", "images")
-
-picture_dimensions = (125, 125)
-
-
-def save_picture(form_picture):
-    random_hex = secrets.token_hex(8)
-    _, file_ext = os.path.splitext(form_picture.filename)
-    profile_filename = random_hex + file_ext
-    picture_path = os.path.join(image_dir, profile_filename)
-
-    with Image.open(form_picture) as i:
-        i.thumbnail(picture_dimensions)
-        i.save(picture_path)
-    return profile_filename
 
 
 @users_blueprint.route("/account", methods=["GET", "POST"])
@@ -43,7 +22,7 @@ def account():
         current_user.email = form.email.data
         db.session.commit()
         flash(message="Your account has been updated!", category="success")
-        return redirect(url_for("account"))
+        return redirect(url_for("users_blueprint.account"))
     elif request.method == "GET":
         form.username.data = current_user.username
         form.email.data = current_user.email
